[Godfile app step 1: extract workflow-task-actions](1) Relocate task/workflow action orchestration
Relocate the GUI task and workflow action orchestration, the auto-fix
observability helpers, and the execution-capacity guard from
packages/app/src/main.ts into packages/app/src/execution/workflow-task-actions.ts,
behind a deps object whose getters resolve the reassignable service singletons.
Re-point the source-pin test anchors that travel with the code. Behavior unchanged.
